% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalcEAD.R
\name{CalcEAD}
\alias{CalcEAD}
\title{Calculates the EAD}
\usage{
CalcEAD(RC, PFE)
}
\arguments{
\item{RC}{the replacement cost}

\item{PFE}{the projected future exposure}
}
\value{
The Exposure-at-Default
}
\description{
Calculates the Exposure at Default
}
\examples{
#returns 1.4*(60+500) = 784
EAD <- CalcEAD(60,500)
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

