% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Commodity.R
\docType{class}
\name{CommSwap-class}
\alias{CommSwap}
\title{Commodity Swap Class}
\value{
An object of type CommSwap
}
\description{
Creates a Commodity Swap Object with the relevant info needed to calculate the Exposure-at-Default (EAD)
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

