% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetermineCCRMethodology.R
\name{DetermineCCRMethodology}
\alias{DetermineCCRMethodology}
\title{Specifies the CCR methodology}
\usage{
DetermineCCRMethodology(trades_filename, total_assets)
}
\arguments{
\item{trades_filename}{the file holding the trades of the portfolio}

\item{total_assets}{the total assets of the institution in mio EUR}
}
\value{
The CCR methodology that the institution is eligible to utilize
}
\description{
Determines the CCR methodology that the institution is eligible to utilize.
The regulator allows the institutions to select less complicated methodologies when the derivatives trading business is negligible
}
\references{
Regulation (EU) 2019/876 of the European Parliament and of the Council of 20 May 2019
http://data.europa.eu/eli/reg/2019/876/oj
}
\author{
Tasos Grivas <info@openriskcalculator.com>
}
