% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleBasisVol.R
\name{ExampleBasisVol}
\alias{ExampleBasisVol}
\title{Basis+Volatility trades Example}
\usage{
ExampleBasisVol(JSON = FALSE)
}
\arguments{
\item{JSON}{(optional) if TRUE it returns a json string}
}
\value{
The exposure at default
}
\description{
Calculates the Exposure at Default for a trade set containing basis and volatility transactions
}
\references{
Regulation (EU) 2019/876 of the European Parliament and of the Council of 20 May 2019
http://data.europa.eu/eli/reg/2019/876/oj
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
