% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExampleCredit.R
\name{ExampleCredit}
\alias{ExampleCredit}
\title{Credit Products Example}
\usage{
ExampleCredit()
}
\value{
The exposure at default (expected value based on the Basel paper is 381)
}
\description{
Calculates the Exposure at Default for the Credit example as given in the Basel III regulatory paper
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

