% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExampleFX.R
\name{ExampleFX}
\alias{ExampleFX}
\title{FX Example}
\usage{
ExampleFX()
}
\value{
The exposure at default (expected value is 364)
}
\description{
Calculates the Exposure at Default for the FX product type
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

