% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleIRD.R
\name{ExampleIRD}
\alias{ExampleIRD}
\title{IRDs Example}
\usage{
ExampleIRD(JSON = FALSE)
}
\arguments{
\item{JSON}{(optional) if TRUE it returns a json string}
}
\value{
The exposure at default (expected value based on the Basel paper is 569)
}
\description{
Calculates the Exposure at Default for the IRD example as given in the Basel III regulatory paper
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
