% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleIRDCredit.R
\name{ExampleIRDCredit}
\alias{ExampleIRDCredit}
\title{IRDs+Commodity Example}
\usage{
ExampleIRDCredit(JSON = FALSE)
}
\arguments{
\item{JSON}{(optional) if TRUE it returns a json string}
}
\value{
The exposure at default (expected value  based on the Basel paper is 936)
}
\description{
Calculates the Exposure at Default for the IRDs + Commodity example as given in the Basel III regulatory paper
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

