% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HandleBasisVol.R
\name{HandleBasisVol}
\alias{HandleBasisVol}
\title{Calculates the Addon amount after handling basis and Volatility trades}
\usage{
HandleBasisVol(trades)
}
\arguments{
\item{trades}{The full list of the Trade Objects}
}
\value{
The aggregate amount of the addon summed up for all the asset classes
}
\description{
Calculates the addon amount after splitting the trades into 'basis swap', 'volatility' and 'normal' transactions.
The corresponding penalty factors are applied to the supervisory factors for each trade group.
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

