% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IRD.R
\docType{class}
\name{IRDSwapVol-class}
\alias{IRDSwapVol}
\title{IRD Swap Volatility Class}
\value{
An object of type IRDSwapVol
}
\description{
Creates an IRD Swap Volatility-based Object with the relevant info needed to calculate the Exposure-at-Default (EAD)
}

