% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalcAddon.R
\name{CalcAddon}
\alias{CalcAddon}
\title{Calculates the Addon amount}
\usage{
CalcAddon(trades, MF, factor_mult)
}
\arguments{
\item{trades}{The full list of the Trade Objects}

\item{MF}{(Optional) The Maturity Factor based on the collateral agreement}

\item{factor_mult}{(Optional) The Multiplication Factor applicable for volatility/basis trades}
}
\value{
The aggregate amount of the addon summed up for all the asset classes
}
\description{
Calculates the aggregate amount of the addon after splitting per asset class
and dividing the trades into the corresponding netting sets per currency, timebucket  etc.
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

