% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPFE.R
\name{CalcPFE}
\alias{CalcPFE}
\title{Calculates the PFE}
\usage{
CalcPFE(V_C, Addon_Aggregate)
}
\arguments{
\item{V_C}{the difference between the sum of the MtMs and the collateral}

\item{Addon_Aggregate}{the aggregate amount of the Addon}
}
\value{
The Projected Future Exposure (PFE)
}
\description{
Calculates the Projected Future Exposure (PFE) after applying the relevant multiplier.
The purpose of this multiplier is to lessen the risk stemming from the addons in case of excess collateral
}
\author{
Project team <info@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

