% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalcRC.R
\name{CalcRC}
\alias{CalcRC}
\title{Calculates the RC}
\usage{
CalcRC(trades)
}
\arguments{
\item{trades}{The full list of the Trade Objects}
}
\value{
The replacement Cost and the sum of the MtMs
}
\description{
Calculates the Replacement Cost(RC) and the sum of the MtMs for all the trades
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

