% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExampleBasisVol.R
\name{ExampleBasisVol}
\alias{ExampleBasisVol}
\title{Basis+Volatility trades Example}
\usage{
ExampleBasisVol()
}
\value{
The exposure at default
}
\description{
Calculates the Exposure at Default for a trade set containing basis and volatility transactions
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

