% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleFX.R
\name{ExampleFX}
\alias{ExampleFX}
\title{FX Example}
\usage{
ExampleFX(JSON = FALSE)
}
\arguments{
\item{JSON}{(optional) if TRUE it returns a json string}
}
\value{
The exposure at default
}
\description{
Calculates the Exposure at Default for the FX product type
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
