% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleBasisVol.R
\name{HandleBasisVol}
\alias{HandleBasisVol}
\title{Splits trades in being basis, volatility or 'normal' transactions}
\usage{
HandleBasisVol(trades)
}
\arguments{
\item{trades}{The full list of the Trade Objects}
}
\value{
A list depicting which trade IDs fall under each hedging set.
}
\description{
Receives a list of trades and splits them according to being basis, volatility or 'normal' transactions
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

