% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleBasisVol.R
\name{HandleBasisVol}
\alias{HandleBasisVol}
\title{Splits trades in being basis, volatility or 'normal' transactions}
\usage{
HandleBasisVol(trades)
}
\arguments{
\item{trades}{The full list of the Trade Objects}
}
\value{
A list depicting which trade IDs fall under each hedging set.
}
\description{
Receives a list of trades and splits them according to being basis, volatility or 'normal' transactions
}
\references{
Regulation (EU) 2019/876 of the European Parliament and of the Council of 20 May 2019
http://data.europa.eu/eli/reg/2019/876/oj
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
