% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSupervisoryData.R
\name{LoadSupervisoryData}
\alias{LoadSupervisoryData}
\title{Supervisory Data Loading}
\usage{
LoadSupervisoryData()
}
\value{
A data frame with the required data
}
\description{
Loads the supervisory data (factors, correlation and option volatility)
for each Asset Class and SubClass
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

