% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACCRCalculator.R
\name{SACCRCalculator}
\alias{SACCRCalculator}
\title{SA-CCR Calculator}
\usage{
SACCRCalculator(trades_filename, csa_filename, coll_filename, JSON = FALSE)
}
\arguments{
\item{trades_filename}{a .csv file containing the trades}

\item{csa_filename}{a .csv file containing CSAs}

\item{coll_filename}{a .csv file containing collaterals}

\item{JSON}{(optional) if TRUE it returns a json string}
}
\value{
The exposure at default (expected value based on the Basel paper is 569)
}
\description{
Returns a tree structure depicting the add-on calculations on different hedging/netting sets
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

