% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTradeAddon.R
\name{SingleTradeAddon}
\alias{SingleTradeAddon}
\title{Calculates the addon information}
\usage{
SingleTradeAddon(trade, MF)
}
\arguments{
\item{trade}{A trade object}

\item{MF}{(Optional) The Maturity Factor based on the collateral agreement}
}
\value{
A list of addon information
}
\description{
Calculates the addon information (including Adj notional, superv delta etc) for each trade
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
\references{
Basel Committee: The standardised approach for measuring counterparty credit risk exposures
http://www.bis.org/publ/bcbs279.htm
}

